CCS881 baseline and cleanup
