--- libraries/SparkFun_CCS811/library.properties
@@ -1,9 +1,0 @@
-name=SparkFun CCS811 Arduino Library
-version=1.0.7
-author=SparkFun Electronics <techsupport@sparkfun.com>
-maintainer=SparkFun Electronics <sparkfun.com>
-sentence=An Arduino library to drive the AMS CCS811 by I2C.
-paragraph=The <a href="https://www.sparkfun.com/products/14193">CCS811 Air Quality Breakout</a> is a digital gas sensor solution that senses a wide range of Total Volatile Organic Compounds (TVOCs), including equivalent carbon dioxide (eCO2) and metal oxide (MOX) levels. It is intended for indoor air quality monitoring in personal devices such as watches and phones, but we’ve put it on a breakout board so you can use it as a regular I2C device.
-category=Sensors
-url=https://github.com/sparkfun/SparkFun_CCS811_Arduino_Library
-architectures=*
